visu/requirements.txt
Pillow
colour